Report only unique error output lines
In the case of ESLint, we are seeing a recursive stack-trace error for a
certain situation. This makes the error output very long and causes us
to send this back and store it on Hound's end. In turn, GitHub limits
the review body comment to 64KB which this error output can go over.