Strip out temp path from error outputs
Some linters will print out the filename of the config file when they
encounter an error with its contents. Since we write the config file
(and run the linters) from a temp path, the output can be sometimes
confusing. Let's strip out the temp path and just report the filename.

The test handles both unix (Heroku) and OSX temp paths.