Fix the test for linter error case
The test was passing because the linter was erroring due to `Enabled`
not being indented properly. It also happened to detect wrong namespace
for `Style/AlignHash`. This wasn't quite the intention of the test.
Thus, I've clarified it by adding another rule that actually causes the
linter to exit with non-zero status, and now test against the exact/full
error output.